Applied brightness, contrast and sharpness effects to the crop editor camera feed
